feat: add Scaffold Item in app, change NuGet version
Added a scaffolded item to the project.
We use the standard registration and login provided by ASP, connect them to the Program, and add them to the Layout.
The nuget package versions were also changed from 6.0.22 - 6.0.21 and 7.0.0 - 6.0.21 due to a compatibility error. (we can see which packages have been modified in webApp.csproj)